TECH-4975: split out foreign_key_settings
